+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added input form for hydrological unit codes
 - added input form for up to 10 counties
 - added longitude and latitude columns to data Schema
+- updated time column in schema to use tableau's datetime format, and created a script to translate into tableau-compliant datetime format. 
 ### Changed